src/voicebot_app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
qdrant-client==1.9.0
sentence-transformers==3.0.1
huggingface-hub==0.23.2
torch==2.1.0
numpy==1.26.4
python-dotenv==1.0.1
starlette==0.37.2
rank-bm25==0.2.2
aiohttp==3.9.5
sounddevice
soundfile
webrtcvad==2.0.10
scipy==1.13.0
deepgram-sdk>=3.0.0